refactor: split 6 large files under 500 lines
Extract shared helpers into internal headers, distribute methods/tests
across focused .cpp files. No behavior changes — pure file-level
decomposition for faster incremental builds.

- telemetry_engine.cpp (994→359): sensors + process split out
- c_api.cpp (831→264): format APIs split out, shared internal header
- test_platform_native.cpp (838→336): config + store tests split out
- cockpit_controller.cpp (643→300): support methods split out
- cockpit_controller_tests.cpp (818→242): sensor tests split out
- dock_model_tests.cpp (732→443): invariant tests split out
