+++ tests/test_platform/native/CMakeLists.txt
@@ -1,4 +1,8 @@
-add_executable(platform_native_tests test_platform_native.cpp)
+add_executable(platform_native_tests
+    test_platform_native.cpp
+    test_platform_config.cpp
+    test_platform_store.cpp
+)
 
 target_include_directories(platform_native_tests
     PRIVATE
