+++ tests/test_shell/native/cockpit_controller_tests.cpp
@@ -1,346 +1,13 @@
+#include "cockpit_test_fakes.hpp"
+#include "cockpit_test_registry.hpp"
+
 #include "aura_shell/cockpit_controller.hpp"
 
-#include <algorithm>
 #include <cmath>
 #include <iostream>
 #include <limits>
 #include <memory>
-#include <optional>
 #include <string>
-#include <vector>
-
-namespace {
-
-class FakeTelemetryBridge final : public aura::shell::ITelemetryBridge {
-public:
-    bool backend_available = true;
-    std::optional<aura::shell::TelemetrySnapshot> next_snapshot{
-        aura::shell::TelemetrySnapshot{35.0, 48.0}
-    };
-    std::vector<aura::shell::ProcessSample> next_processes{
-        aura::shell::ProcessSample{1234U, "aura", 22.1, 32ULL * 1024ULL * 1024ULL},
-        aura::shell::ProcessSample{2048U, "explorer", 7.5, 110ULL * 1024ULL * 1024ULL},
-    };
-    std::string snapshot_error;
-    std::string process_error;
-
-    bool available() const override {
-        return backend_available;
-    }
-
-    std::optional<aura::shell::TelemetrySnapshot> collect_snapshot(std::string& error) override {
-        if (!backend_available || !next_snapshot.has_value()) {
-            error = snapshot_error.empty() ? "telemetry unavailable" : snapshot_error;
-            return std::nullopt;
-        }
-        error.clear();
-        return next_snapshot;
-    }
-
-    std::vector<aura::shell::ProcessSample> collect_top_processes(
-        const std::size_t max_samples,
-        std::string& error
-    ) override {
-        if (!backend_available) {
-            error = process_error.empty() ? "telemetry unavailable" : process_error;
-            return {};
-        }
-        error = process_error;
-        std::vector<aura::shell::ProcessSample> output = next_processes;
-        if (output.size() > max_samples) {
-            output.resize(max_samples);
-        }
-        return output;
-    }
-
-    // ── Extended sensor fakes ────────────────────────────────────────────
-    std::optional<aura::shell::PerCoreCpuState> next_per_core;
-    std::optional<aura::shell::GpuState> next_gpu;
-    std::optional<aura::shell::DiskIoState> next_disk;
-    std::optional<aura::shell::NetworkIoState> next_network;
-    std::optional<aura::shell::ThermalState> next_thermal;
-
-    std::optional<aura::shell::PerCoreCpuState> collect_per_core_cpu(std::string& error) override {
-        error.clear();
-        if (!backend_available) { error = "unavailable"; return std::nullopt; }
-        return next_per_core;
-    }
-    std::optional<aura::shell::GpuState> collect_gpu(std::string& error) override {
-        error.clear();
-        if (!backend_available) { error = "unavailable"; return std::nullopt; }
-        return next_gpu;
-    }
-    std::optional<aura::shell::DiskIoState> collect_disk_io(std::string& error) override {
-        error.clear();
-        if (!backend_available) { error = "unavailable"; return std::nullopt; }
-        return next_disk;
-    }
-    std::optional<aura::shell::NetworkIoState> collect_network_io(std::string& error) override {
-        error.clear();
-        if (!backend_available) { error = "unavailable"; return std::nullopt; }
-        return next_network;
-    }
-    std::optional<aura::shell::ThermalState> collect_thermal(std::string& error) override {
-        error.clear();
-        if (!backend_available) { error = "unavailable"; return std::nullopt; }
-        return next_thermal;
-    }
-    std::vector<aura::shell::ProcessSample> collect_process_details(
-        std::size_t max_results, std::uint8_t /*sort_column*/,
-        bool /*sort_descending*/, std::string& error) override {
-        return collect_top_processes(max_results, error);
-    }
-    bool terminate_process(std::uint32_t /*pid*/, std::string& error) override {
-        error = "not implemented in fake";
-        return false;
-    }
-};
-
-class FakeRenderBridge final : public aura::shell::IRenderBridge {
-public:
-    bool backend_available = true;
-    bool fail_compose = false;
-    bool fail_style_tokens = false;
-    bool fail_lines = false;
-    bool fail_rows = false;
-    bool fail_status = false;
-
-    bool available() const override {
-        return backend_available;
-    }
-
-    double sanitize_percent(const double value) const override {
-        if (!std::isfinite(value)) {
-            return 0.0;
-        }
-        return std::clamp(value, 0.0, 100.0);
-    }
-
-    std::optional<aura::shell::FrameState> compose_frame(
-        const double previous_phase,
-        const double elapsed_since_last_frame,
-        const double cpu_percent,
-        const double memory_percent,
-        std::string& error
-    ) const override {
-        if (!backend_available || fail_compose) {
-            error = "compose failed";
-            return std::nullopt;
-        }
-        error.clear();
-        aura::shell::FrameState frame;
-        frame.phase = std::fmod(previous_phase + std::max(0.0, elapsed_since_last_frame), 1.0);
-        frame.accent_intensity = std::clamp((cpu_percent + memory_percent) / 200.0, 0.0, 1.0);
-        frame.next_delay_seconds = 1.0 / 60.0;
-        return frame;
-    }
-
-    std::optional<aura::shell::RenderStyleTokens> compute_style_tokens(
-        const double previous_phase,
-        const double elapsed_since_last_frame,
-        const double cpu_percent,
-        const double memory_percent,
-        std::string& error
-    ) const override {
-        if (!backend_available || fail_style_tokens) {
-            error = "style tokens failed";
-            return std::nullopt;
-        }
-        error.clear();
-        aura::shell::RenderStyleTokens tokens;
-        tokens.phase = std::fmod(previous_phase + std::max(0.0, elapsed_since_last_frame), 1.0);
-        tokens.next_delay_seconds = 1.0 / 60.0;
-        tokens.accent_intensity = std::clamp((cpu_percent + memory_percent) / 200.0, 0.0, 1.0);
-        tokens.accent_red = std::clamp(0.2 + tokens.accent_intensity * 0.4, 0.0, 1.0);
-        tokens.accent_green = std::clamp(0.4 + tokens.accent_intensity * 0.3, 0.0, 1.0);
-        tokens.accent_blue = std::clamp(0.7 + tokens.accent_intensity * 0.2, 0.0, 1.0);
-        tokens.accent_alpha = std::clamp(0.2 + tokens.accent_intensity * 0.4, 0.0, 1.0);
-        tokens.frost_intensity = std::clamp(0.2 + tokens.accent_intensity * 0.7, 0.0, 1.0);
-        tokens.tint_strength = std::clamp(0.3 + tokens.accent_intensity * 0.5, 0.0, 1.0);
-        tokens.ring_line_width = std::clamp(1.0 + tokens.accent_intensity * 6.0, 1.0, 7.0);
-        tokens.ring_glow_strength = tokens.accent_intensity;
-        tokens.cpu_alpha = std::clamp(0.3 + sanitize_percent(cpu_percent) / 100.0 * 0.7, 0.0, 1.0);
-        tokens.memory_alpha = std::clamp(0.3 + sanitize_percent(memory_percent) / 100.0 * 0.7, 0.0, 1.0);
-        tokens.severity_level = tokens.accent_intensity > 0.8 ? 3 : (tokens.accent_intensity > 0.6 ? 2 : 1);
-        tokens.motion_scale = tokens.severity_level >= 2 ? 0.75 : 0.95;
-        tokens.quality_hint = tokens.severity_level >= 2 ? 1 : 0;
-        tokens.timeline_anomaly_alpha = std::clamp(tokens.accent_intensity, 0.05, 1.0);
-        return tokens;
-    }
-
-    std::optional<aura::shell::SnapshotLines> format_snapshot_lines(
-        const double timestamp,
-        const double cpu_percent,
-        const double memory_percent,
-        std::string& error
-    ) const override {
-        if (!backend_available || fail_lines) {
-            error = "snapshot formatting failed";
-            return std::nullopt;
-        }
-        error.clear();
-        aura::shell::SnapshotLines lines;
-        lines.cpu = "cpu_line_" + std::to_string(static_cast<int>(cpu_percent));
-        lines.memory = "mem_line_" + std::to_string(static_cast<int>(memory_percent));
-        lines.timestamp = "ts_line_" + std::to_string(static_cast<int>(timestamp));
-        return lines;
-    }
-
-    std::optional<std::string> format_process_row(
-        const int rank,
-        const std::string& name,
-        const double cpu_percent,
-        const double memory_rss_bytes,
-        const int /*max_chars*/,
-        std::string& error
-    ) const override {
-        if (!backend_available || fail_rows) {
-            error = "row formatting failed";
-            return std::nullopt;
-        }
-        error.clear();
-        return "#" + std::to_string(rank) + " " + name + " cpu=" + std::to_string(static_cast<int>(cpu_percent)) +
-               " mem=" + std::to_string(static_cast<int>(memory_rss_bytes));
-    }
-
-    std::optional<std::string> format_stream_status(
-        const std::optional<std::string>& db_path,
-        const std::optional<int>& /*sample_count*/,
-        const std::optional<std::string>& stream_error,
-        std::string& error
-    ) const override {
-        if (!backend_available || fail_status) {
-            error = "status formatting failed";
-            return std::nullopt;
-        }
-        error.clear();
-        std::string value = "db=" + (db_path.has_value() ? *db_path : std::string("<none>")) + " render=ok";
-        if (stream_error.has_value() && !stream_error->empty()) {
-            value += " warning=" + *stream_error;
-        }
-        return value;
-    }
-
-    std::string last_error_text() const override {
-        if (!backend_available) {
-            return "render unavailable";
-        }
-        if (fail_style_tokens) {
-            return "style tokens failed";
-        }
-        return {};
-    }
-};
-
-class FakeTimelineBridge final : public aura::shell::ITimelineBridge {
-public:
-    bool backend_available = true;
-    bool fail_query = false;
-    std::string query_error;
-    std::vector<aura::shell::TimelinePoint> next_points{
-        {1699999950.0, 20.0, 35.0},
-        {1699999955.0, 21.0, 35.5},
-        {1699999960.0, 22.0, 36.0},
-        {1699999965.0, 23.0, 36.5},
-        {1699999970.0, 24.0, 37.0},
-        {1699999975.0, 25.0, 37.5},
-        {1699999980.0, 26.0, 38.0},
-        {1699999985.0, 27.0, 38.5},
-        {1699999990.0, 28.0, 39.0},
-        {1699999995.0, 29.0, 39.5},
-    };
-    int query_count = 0;
-
-    bool available() const override {
-        return backend_available;
-    }
-
-    std::vector<aura::shell::TimelinePoint> query_recent(
-        const std::string& db_path,
-        const double /*end_timestamp*/,
-        const double /*window_seconds*/,
-        const int resolution,
-        std::string& error
-    ) override {
-        ++query_count;
-        if (!backend_available || fail_query) {
-            error = query_error.empty() ? "timeline unavailable" : query_error;
-            return {};
-        }
-        if (db_path.empty()) {
-            error = "db_path empty";
-            return {};
-        }
-        error.clear();
-        std::vector<aura::shell::TimelinePoint> output = next_points;
-        if (resolution > 0 && output.size() > static_cast<std::size_t>(resolution)) {
-            output.resize(static_cast<std::size_t>(resolution));
-        }
-        return output;
-    }
-};
-
-bool expect_true(const bool condition, const std::string& name) {
-    if (!condition) {
-        std::cerr << "FAILED: " << name << '\n';
-        return false;
-    }
-    return true;
-}
-
-bool contains(const std::string& text, const std::string& pattern) {
-    return text.find(pattern) != std::string::npos;
-}
-
-bool style_tokens_valid(const aura::shell::RenderStyleTokens& tokens) {
-    return std::isfinite(tokens.phase) &&
-           std::isfinite(tokens.next_delay_seconds) &&
-           std::isfinite(tokens.accent_intensity) &&
-           std::isfinite(tokens.accent_red) &&
-           std::isfinite(tokens.accent_green) &&
-           std::isfinite(tokens.accent_blue) &&
-           std::isfinite(tokens.accent_alpha) &&
-           std::isfinite(tokens.frost_intensity) &&
-           std::isfinite(tokens.tint_strength) &&
-           std::isfinite(tokens.ring_line_width) &&
-           std::isfinite(tokens.ring_glow_strength) &&
-           std::isfinite(tokens.cpu_alpha) &&
-           std::isfinite(tokens.memory_alpha) &&
-           std::isfinite(tokens.motion_scale) &&
-           std::isfinite(tokens.timeline_anomaly_alpha) &&
-           tokens.phase >= 0.0 &&
-           tokens.phase < 1.0 &&
-           tokens.next_delay_seconds >= 0.0 &&
-           tokens.accent_intensity >= 0.0 &&
-           tokens.accent_intensity <= 1.0 &&
-           tokens.accent_red >= 0.0 &&
-           tokens.accent_red <= 1.0 &&
-           tokens.accent_green >= 0.0 &&
-           tokens.accent_green <= 1.0 &&
-           tokens.accent_blue >= 0.0 &&
-           tokens.accent_blue <= 1.0 &&
-           tokens.accent_alpha >= 0.0 &&
-           tokens.accent_alpha <= 1.0 &&
-           tokens.frost_intensity >= 0.0 &&
-           tokens.frost_intensity <= 1.0 &&
-           tokens.tint_strength >= 0.0 &&
-           tokens.tint_strength <= 1.0 &&
-           tokens.ring_line_width >= 1.0 &&
-           tokens.ring_line_width <= 7.0 &&
-           tokens.ring_glow_strength >= 0.0 &&
-           tokens.ring_glow_strength <= 1.0 &&
-           tokens.cpu_alpha >= 0.0 &&
-           tokens.cpu_alpha <= 1.0 &&
-           tokens.memory_alpha >= 0.0 &&
-           tokens.memory_alpha <= 1.0 &&
-           tokens.severity_level >= 0 &&
-           tokens.severity_level <= 3 &&
-           tokens.motion_scale >= 0.60 &&
-           tokens.motion_scale <= 1.0 &&
-           (tokens.quality_hint == 0 || tokens.quality_hint == 1) &&
-           tokens.timeline_anomaly_alpha >= 0.0 &&
-           tokens.timeline_anomaly_alpha <= 1.0;
-}
 
 bool test_happy_path_prefers_dvr() {
     auto telemetry = std::make_unique<FakeTelemetryBridge>();
@@ -514,255 +181,10 @@ bool test_last_good_reused_on_telemetry_failure_preserves_timeline() {
     return ok;
 }
 
-bool test_falls_back_to_live_when_dvr_unavailable() {
-    auto telemetry = std::make_unique<FakeTelemetryBridge>();
-    auto render = std::make_unique<FakeRenderBridge>();
-    auto timeline = std::make_unique<FakeTimelineBridge>();
-    timeline->backend_available = false;
-
-    aura::shell::CockpitController::Config config;
-    config.db_path = "C:/tmp/aura.db";
-
-    aura::shell::CockpitController controller(
-        std::move(telemetry),
-        std::move(render),
-        std::move(timeline),
-        config
-    );
-    static_cast<void>(controller.tick(1.0, 1700000010.0));
-    const aura::shell::CockpitUiState state = controller.tick(1.0, 1700000011.0);
-
-    bool ok = true;
-    ok &= expect_true(state.timeline_source == aura::shell::TimelineSource::Live, "fallback-live: source");
-    ok &= expect_true(state.timeline_points.size() >= 2U, "fallback-live: points");
-    ok &= expect_true(contains(state.timeline_line, "timeline=live"), "fallback-live: line");
-    return ok;
-}
-
-bool test_live_ring_respects_capacity() {
-    auto telemetry = std::make_unique<FakeTelemetryBridge>();
-    auto* telemetry_ptr = telemetry.get();
-    auto render = std::make_unique<FakeRenderBridge>();
-    auto timeline = std::make_unique<FakeTimelineBridge>();
-
-    aura::shell::CockpitController::Config config;
-    config.prefer_dvr_timeline = false;
-    config.timeline_live_capacity = 3U;
-    config.timeline_window_seconds = 1000.0;
-
-    aura::shell::CockpitController controller(
-        std::move(telemetry),
-        std::move(render),
-        std::move(timeline),
-        config
-    );
-
-    aura::shell::CockpitUiState state;
-    for (int i = 0; i < 5; ++i) {
-        telemetry_ptr->next_snapshot = aura::shell::TelemetrySnapshot{
-            10.0 + static_cast<double>(i),
-            20.0 + static_cast<double>(i),
-        };
-        state = controller.tick(1.0, 1700000100.0 + static_cast<double>(i));
-    }
-
-    bool ok = true;
-    ok &= expect_true(state.timeline_source == aura::shell::TimelineSource::Live, "capacity: source");
-    ok &= expect_true(state.timeline_points.size() == 3U, "capacity: size");
-    ok &= expect_true(std::fabs(state.timeline_points.front().timestamp - 1700000102.0) < 0.0001, "capacity: front timestamp");
-    ok &= expect_true(state.timeline_anomaly_count >= 0, "capacity: anomaly count present");
-    return ok;
-}
-
-bool test_anomaly_count_detects_spikes() {
-    auto telemetry = std::make_unique<FakeTelemetryBridge>();
-    auto* telemetry_ptr = telemetry.get();
-    auto render = std::make_unique<FakeRenderBridge>();
-    auto timeline = std::make_unique<FakeTimelineBridge>();
-
-    aura::shell::CockpitController::Config config;
-    config.prefer_dvr_timeline = false;
-    config.timeline_live_capacity = 8U;
-
-    aura::shell::CockpitController controller(
-        std::move(telemetry),
-        std::move(render),
-        std::move(timeline),
-        config
-    );
-
-    static_cast<void>(controller.tick(1.0, 1700000200.0));
-    telemetry_ptr->next_snapshot = aura::shell::TelemetrySnapshot{92.0, 91.0};
-    const aura::shell::CockpitUiState state = controller.tick(1.0, 1700000201.0);
-
-    bool ok = true;
-    ok &= expect_true(state.timeline_source == aura::shell::TimelineSource::Live, "anomaly: live source");
-    ok &= expect_true(state.timeline_anomaly_count >= 1, "anomaly: count spike");
-    ok &= expect_true(contains(state.timeline_line, "anomalies="), "anomaly: line contains anomalies");
-    return ok;
-}
-
-bool test_per_core_cpu_flows_through() {
-    auto telemetry = std::make_unique<FakeTelemetryBridge>();
-    aura::shell::PerCoreCpuState per_core;
-    per_core.core_count = 4;
-    per_core.core_percents = {25.0, 50.0, 75.0, 100.0};
-    telemetry->next_per_core = per_core;
-    auto render = std::make_unique<FakeRenderBridge>();
-    auto timeline = std::make_unique<FakeTimelineBridge>();
-
-    aura::shell::CockpitController controller(
-        std::move(telemetry), std::move(render), std::move(timeline), {}
-    );
-    const auto state = controller.tick(1.0, 1700001000.0);
-
-    bool ok = true;
-    ok &= expect_true(state.per_core_cpu.core_count == 4, "per-core: count");
-    ok &= expect_true(state.per_core_cpu.core_percents.size() == 4, "per-core: vector size");
-    ok &= expect_true(std::fabs(state.per_core_cpu.core_percents[2] - 75.0) < 0.01, "per-core: value");
-    return ok;
-}
-
-bool test_gpu_unavailable_default() {
-    auto telemetry = std::make_unique<FakeTelemetryBridge>();
-    // next_gpu left as nullopt — GPU not available
-    auto render = std::make_unique<FakeRenderBridge>();
-    auto timeline = std::make_unique<FakeTimelineBridge>();
-
-    aura::shell::CockpitController controller(
-        std::move(telemetry), std::move(render), std::move(timeline), {}
-    );
-    const auto state = controller.tick(1.0, 1700001001.0);
-
-    bool ok = true;
-    ok &= expect_true(!state.gpu.available, "gpu-unavail: not available");
-    ok &= expect_true(state.gpu.gpu_percent == 0.0, "gpu-unavail: zero percent");
-    return ok;
-}
-
-bool test_disk_network_rate_flows_through() {
-    auto telemetry = std::make_unique<FakeTelemetryBridge>();
-    aura::shell::DiskIoState disk;
-    disk.read_bytes_per_sec = 1000000.0;
-    disk.write_bytes_per_sec = 500000.0;
-    telemetry->next_disk = disk;
-    aura::shell::NetworkIoState net;
-    net.recv_bytes_per_sec = 2000000.0;
-    net.sent_bytes_per_sec = 100000.0;
-    telemetry->next_network = net;
-    auto render = std::make_unique<FakeRenderBridge>();
-    auto timeline = std::make_unique<FakeTimelineBridge>();
-
-    aura::shell::CockpitController controller(
-        std::move(telemetry), std::move(render), std::move(timeline), {}
-    );
-    const auto state = controller.tick(1.0, 1700001002.0);
-
-    bool ok = true;
-    ok &= expect_true(std::fabs(state.disk_io.read_bytes_per_sec - 1000000.0) < 0.01, "disk: read rate");
-    ok &= expect_true(std::fabs(state.network_io.recv_bytes_per_sec - 2000000.0) < 0.01, "net: recv rate");
-    return ok;
-}
-
-bool test_tiered_polling_gpu_every_2_ticks() {
-    auto telemetry = std::make_unique<FakeTelemetryBridge>();
-    auto* telemetry_ptr = telemetry.get();
-    aura::shell::GpuState gpu;
-    gpu.available = true;
-    gpu.gpu_percent = 42.0;
-    telemetry_ptr->next_gpu = gpu;
-    auto render = std::make_unique<FakeRenderBridge>();
-    auto timeline = std::make_unique<FakeTimelineBridge>();
-
-    aura::shell::CockpitController controller(
-        std::move(telemetry), std::move(render), std::move(timeline), {}
-    );
-
-    // Tick 1: GPU should be collected (first tick)
-    const auto state1 = controller.tick(1.0, 1700002000.0);
-    bool ok = true;
-    ok &= expect_true(state1.gpu.available, "tiered-gpu: available tick 1");
-    ok &= expect_true(std::fabs(state1.gpu.gpu_percent - 42.0) < 0.01, "tiered-gpu: value tick 1");
-
-    // Tick 2: change GPU value — tick_count_=2 (even), should collect
-    gpu.gpu_percent = 88.0;
-    telemetry_ptr->next_gpu = gpu;
-    const auto state2 = controller.tick(1.0, 1700002001.0);
-    ok &= expect_true(std::fabs(state2.gpu.gpu_percent - 88.0) < 0.01, "tiered-gpu: value tick 2");
-
-    // Tick 3: change GPU value — tick_count_=3 (odd), should use cached
-    gpu.gpu_percent = 99.0;
-    telemetry_ptr->next_gpu = gpu;
-    const auto state3 = controller.tick(1.0, 1700002002.0);
-    ok &= expect_true(std::fabs(state3.gpu.gpu_percent - 88.0) < 0.01, "tiered-gpu: cached tick 3");
-
-    return ok;
-}
-
-bool test_graceful_degradation_new_sensors() {
-    auto telemetry = std::make_unique<FakeTelemetryBridge>();
-    // All extended sensors left as nullopt
-    auto render = std::make_unique<FakeRenderBridge>();
-    auto timeline = std::make_unique<FakeTimelineBridge>();
-
-    aura::shell::CockpitController controller(
-        std::move(telemetry), std::move(render), std::move(timeline), {}
-    );
-    const auto state = controller.tick(1.0, 1700003000.0);
-
-    bool ok = true;
-    ok &= expect_true(state.telemetry_available, "graceful: telemetry still available");
-    ok &= expect_true(!state.gpu.available, "graceful: gpu default");
-    ok &= expect_true(state.per_core_cpu.core_count == 0, "graceful: no cores");
-    ok &= expect_true(!state.thermal.available, "graceful: no thermal");
-    ok &= expect_true(state.disk_io.read_bytes_per_sec == 0.0, "graceful: zero disk");
-    ok &= expect_true(state.network_io.recv_bytes_per_sec == 0.0, "graceful: zero net");
-    return ok;
-}
-
-bool test_thermal_every_5_ticks() {
-    auto telemetry = std::make_unique<FakeTelemetryBridge>();
-    auto* telemetry_ptr = telemetry.get();
-    aura::shell::ThermalState thermal;
-    thermal.available = true;
-    thermal.hottest_celsius = 65.0;
-    aura::shell::ThermalSensorReading reading;
-    reading.label = "CPU Package";
-    reading.current_celsius = 65.0;
-    thermal.sensors.push_back(reading);
-    telemetry_ptr->next_thermal = thermal;
-    auto render = std::make_unique<FakeRenderBridge>();
-    auto timeline = std::make_unique<FakeTimelineBridge>();
-
-    aura::shell::CockpitController controller(
-        std::move(telemetry), std::move(render), std::move(timeline), {}
-    );
-
-    // Tick 1: should collect thermal (first tick)
-    const auto state1 = controller.tick(1.0, 1700004000.0);
-    bool ok = true;
-    ok &= expect_true(state1.thermal.available, "thermal-tier: available tick 1");
-
-    // Ticks 2-4: change thermal — should use cached (not divisible by 5)
-    thermal.hottest_celsius = 99.0;
-    telemetry_ptr->next_thermal = thermal;
-    aura::shell::CockpitUiState state_mid;
-    for (int i = 2; i <= 4; ++i) {
-        state_mid = controller.tick(1.0, 1700004000.0 + i);
-    }
-    ok &= expect_true(std::fabs(state_mid.thermal.hottest_celsius - 65.0) < 0.01, "thermal-tier: cached ticks 2-4");
-
-    // Tick 5: should collect thermal (tick_count_ % 5 == 0)
-    const auto state5 = controller.tick(1.0, 1700004005.0);
-    ok &= expect_true(std::fabs(state5.thermal.hottest_celsius - 99.0) < 0.01, "thermal-tier: updated tick 5");
-    return ok;
-}
-
-}  // namespace
-
 int main() {
     int failures = 0;
 
+    // --- Core tests (defined here) ---
     if (!test_happy_path_prefers_dvr()) {
         ++failures;
     }
@@ -781,6 +203,8 @@ int main() {
     if (!test_last_good_reused_on_telemetry_failure_preserves_timeline()) {
         ++failures;
     }
+
+    // --- Sensor tests (defined in cockpit_controller_sensor_tests.cpp) ---
     if (!test_falls_back_to_live_when_dvr_unavailable()) {
         ++failures;
     }
